// database/prisma-thinkthank/generated/zod/outputTypeSchemas/AccountFindManyArgsSchema.ts
import { z } from 'zod'
import type { Prisma } from '@prisma/client'
import { AccountIncludeSchema } from '../inputTypeSchemas/AccountIncludeSchema'
import { AccountWhereInputSchema } from '../inputTypeSchemas/AccountWhereInputSchema'
import { AccountOrderByWithRelationInputSchema } from '../inputTypeSchemas/AccountOrderByWithRelationInputSchema'
import { AccountWhereUniqueInputSchema } from '../inputTypeSchemas/AccountWhereUniqueInputSchema'
import { AccountScalarFieldEnumSchema } from '../inputTypeSchemas/AccountScalarFieldEnumSchema'
import { UserArgsSchema } from '../outputTypeSchemas/UserArgsSchema'
// Select schema needs to be in file to prevent circular imports
//------------------------------------------------------

export const AccountSelectSchema: z.ZodType<Prisma.AccountSelect> = z
    .object({
        id: z.boolean().optional(),
        userId: z.boolean().optional(),
        type: z.boolean().optional(),
        provider: z.boolean().optional(),
        providerAccountId: z.boolean().optional(),
        refresh_token: z.boolean().optional(),
        refresh_token_expires_in: z.boolean().optional(),
        access_token: z.boolean().optional(),
        expires_at: z.boolean().optional(),
        token_type: z.boolean().optional(),
        scope: z.boolean().optional(),
        id_token: z.boolean().optional(),
        session_state: z.boolean().optional(),
        user: z.union([z.boolean(), z.lazy(() => UserArgsSchema)]).optional()
    })
    .strict()

export const AccountFindManyArgsSchema: z.ZodType<Prisma.AccountFindManyArgs> = z
    .object({
        select: AccountSelectSchema.optional(),
        include: AccountIncludeSchema.optional(),
        where: AccountWhereInputSchema.optional(),
        orderBy: z.union([AccountOrderByWithRelationInputSchema.array(), AccountOrderByWithRelationInputSchema]).optional(),
        cursor: AccountWhereUniqueInputSchema.optional(),
        take: z.number().optional(),
        skip: z.number().optional(),
        distinct: z.union([AccountScalarFieldEnumSchema, AccountScalarFieldEnumSchema.array()]).optional()
    })
    .strict()

export default AccountFindManyArgsSchema